Updated field history to include current time
Field

  - changed Field::values(...,time) to  Field::values_at(...,time)
  - changed values_at() to template
  - updated history_time_ to include current time
  - tested and debugged values_at()

Parameters

  - changed Method : <method> :dt_ratio_[min|max] to integers
    interpret as minimum number of sub-cycles (default 1) allowed
    and maximum number of super-cycles (default 1) allowed

MethodGravity

  - added test code for Field::values_at()
  - replaced field.scale with explicit loop

+++ src/Cello/parameters_Config.hpp
@@ -488,8 +488,8 @@ public: // attributes
   int                        num_method;
   double                     method_courant_global;
   std::vector<std::string>   method_list;
-  std::vector<double>        method_dt_ratio_min;
-  std::vector<double>        method_dt_ratio_max;
+  std::vector<int>           method_dt_ratio_min;
+  std::vector<int>           method_dt_ratio_max;
 
   std::vector<int>           method_schedule_index;
